+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Changelog Format
 - [KF] [#14](https://github.com/Ateneo-MISA/MISA-Website/issues/14) about us(feature): fix hexagon image sizing
 - [KF] [#12](https://github.com/Ateneo-MISA/MISA-Website/issues/12) about us(feature): add hero frame and mission/vision frame
 - [KF] [#9](https://github.com/Ateneo-MISA/MISA-Website/issues/9) app(feature): add tailwind css
+- [KF MD] [#5](https://github.com/Ateneo-MISA/MISA-Website/issues/5) home page(feature): add events frame
 - [MF] [#4](https://github.com/Ateneo-MISA/MISA-Website/issues/4) home page(feature): add hero
 - [KF] [#2](https://github.com/Ateneo-MISA/MISA-Website/issues/2) home page(feature): add navbar
 - [KF] [#1](https://github.com/Ateneo-MISA/MISA-Website/issues/1) app(feature): add initial project files